@@ -13,14 +13,14 @@
         }
     ],
     "require": {
-        "maxmind-db/reader": "~1.8",
-        "maxmind/web-service-common": "~0.8",
-        "php": ">=7.2",
+        "maxmind-db/reader": "^1.12.1",
+        "maxmind/web-service-common": "~0.10",
+        "php": ">=8.1",
         "ext-json": "*"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "3.*",
-        "phpunit/phpunit": "^8.0 || ^9.0",
+        "phpunit/phpunit": "^10.0",
         "squizlabs/php_codesniffer": "3.*",
         "phpstan/phpstan": "*"
     },
