@@ -12,17 +12,15 @@ class InvalidRequestException extends HttpException
 {
     /**
      * The code returned by the MaxMind web service.
-     *
-     * @var string
      */
-    public $error;
+    public string $error;
 
     public function __construct(
         string $message,
         string $error,
         int $httpStatus,
         string $uri,
-        \Exception $previous = null
+        ?\Exception $previous = null
     ) {
         $this->error = $error;
         parent::__construct($message, $httpStatus, $uri, $previous);
