@@ -8,84 +8,93 @@ use GeoIp2\Util;
 
 /**
  * This class provides the GeoIP2 Anonymous IP model.
- *
- * @property-read bool $isAnonymous This is true if the IP address belongs to
- *     any sort of anonymous network.
- * @property-read bool $isAnonymousVpn This is true if the IP address is
- *     registered to an anonymous VPN provider. If a VPN provider does not
- *     register subnets under names associated with them, we will likely only
- *     flag their IP ranges using the isHostingProvider property.
- * @property-read bool $isHostingProvider This is true if the IP address belongs
- *     to a hosting or VPN provider (see description of isAnonymousVpn property).
- * @property-read bool $isPublicProxy This is true if the IP address belongs to
- *     a public proxy.
- * @property-read bool $isResidentialProxy This is true if the IP address is
- *     on a suspected anonymizing network and belongs to a residential ISP.
- * @property-read bool $isTorExitNode This is true if the IP address is a Tor
- *     exit node.
- * @property-read string $ipAddress The IP address that the data in the model is
- *     for.
- * @property-read string $network The network in CIDR notation associated with
- *      the record. In particular, this is the largest network where all of the
- *      fields besides $ipAddress have the same value.
  */
-class AnonymousIp extends AbstractModel
+class AnonymousIp implements \JsonSerializable
 {
     /**
-     * @var bool
+     * @var bool this is true if the IP address belongs to
+     *           any sort of anonymous network
      */
-    protected $isAnonymous;
+    public readonly bool $isAnonymous;
 
     /**
-     * @var bool
+     * @var bool This is true if the IP address is
+     *           registered to an anonymous VPN provider. If a VPN provider does not
+     *           register subnets under names associated with them, we will likely only
+     *           flag their IP ranges using the isHostingProvider property.
      */
-    protected $isAnonymousVpn;
+    public readonly bool $isAnonymousVpn;
 
     /**
-     * @var bool
+     * @var bool this is true if the IP address belongs
+     *           to a hosting or VPN provider (see description of isAnonymousVpn property)
      */
-    protected $isHostingProvider;
+    public readonly bool $isHostingProvider;
 
     /**
-     * @var bool
+     * @var bool this is true if the IP address belongs to
+     *           a public proxy
      */
-    protected $isPublicProxy;
+    public readonly bool $isPublicProxy;
 
     /**
-     * @var bool
+     * @var bool this is true if the IP address is
+     *           on a suspected anonymizing network and belongs to a residential ISP
      */
-    protected $isResidentialProxy;
+    public readonly bool $isResidentialProxy;
 
     /**
-     * @var bool
+     * @var bool this is true if the IP address is a Tor
+     *           exit node
      */
-    protected $isTorExitNode;
+    public readonly bool $isTorExitNode;
 
     /**
-     * @var string
+     * @var string the IP address that the data in the model is
+     *             for
      */
-    protected $ipAddress;
+    public readonly string $ipAddress;
 
     /**
-     * @var string
+     * @var string The network in CIDR notation associated with
+     *             the record. In particular, this is the largest network where all of the
+     *             fields besides $ipAddress have the same value.
      */
-    protected $network;
+    public readonly string $network;
 
     /**
      * @ignore
+     *
+     * @param array<string, mixed> $raw
      */
     public function __construct(array $raw)
     {
-        parent::__construct($raw);
-
-        $this->isAnonymous = $this->get('is_anonymous');
-        $this->isAnonymousVpn = $this->get('is_anonymous_vpn');
-        $this->isHostingProvider = $this->get('is_hosting_provider');
-        $this->isPublicProxy = $this->get('is_public_proxy');
-        $this->isResidentialProxy = $this->get('is_residential_proxy');
-        $this->isTorExitNode = $this->get('is_tor_exit_node');
-        $ipAddress = $this->get('ip_address');
+        $this->isAnonymous = $raw['is_anonymous'] ?? false;
+        $this->isAnonymousVpn = $raw['is_anonymous_vpn'] ?? false;
+        $this->isHostingProvider = $raw['is_hosting_provider'] ?? false;
+        $this->isPublicProxy = $raw['is_public_proxy'] ?? false;
+        $this->isResidentialProxy = $raw['is_residential_proxy'] ?? false;
+        $this->isTorExitNode = $raw['is_tor_exit_node'] ?? false;
+        $ipAddress = $raw['ip_address'];
         $this->ipAddress = $ipAddress;
-        $this->network = Util::cidr($ipAddress, $this->get('prefix_len'));
+        $this->network = Util::cidr($ipAddress, $raw['prefix_len']);
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    public function jsonSerialize(): ?array
+    {
+        $js = [];
+        $js['is_anonymous'] = $this->isAnonymous;
+        $js['is_anonymous_vpn'] = $this->isAnonymousVpn;
+        $js['is_hosting_provider'] = $this->isHostingProvider;
+        $js['is_public_proxy'] = $this->isPublicProxy;
+        $js['is_residential_proxy'] = $this->isResidentialProxy;
+        $js['is_tor_exit_node'] = $this->isTorExitNode;
+        $js['ip_address'] = $this->ipAddress;
+        $js['network'] = $this->network;
+
+        return $js;
     }
 }
