@@ -8,51 +8,66 @@ use GeoIp2\Util;
 
 /**
  * This class provides the GeoLite2 ASN model.
- *
- * @property-read int|null $autonomousSystemNumber The autonomous system number
- *     associated with the IP address.
- * @property-read string|null $autonomousSystemOrganization The organization
- *     associated with the registered autonomous system number for the IP
- *     address.
- * @property-read string $ipAddress The IP address that the data in the model is
- *     for.
- * @property-read string $network The network in CIDR notation associated with
- *      the record. In particular, this is the largest network where all of the
- *      fields besides $ipAddress have the same value.
  */
-class Asn extends AbstractModel
+class Asn implements \JsonSerializable
 {
     /**
-     * @var int|null
+     * @var int|null the autonomous system number
+     *               associated with the IP address
      */
-    protected $autonomousSystemNumber;
+    public readonly ?int $autonomousSystemNumber;
 
     /**
-     * @var string|null
+     * @var string|null the organization
+     *                  associated with the registered autonomous system number for the IP
+     *                  address
      */
-    protected $autonomousSystemOrganization;
+    public readonly ?string $autonomousSystemOrganization;
 
     /**
-     * @var string
+     * @var string the IP address that the data in the model is
+     *             for
      */
-    protected $ipAddress;
+    public readonly string $ipAddress;
 
     /**
-     * @var string
+     * @var string The network in CIDR notation associated with
+     *             the record. In particular, this is the largest network where all of the
+     *             fields besides $ipAddress have the same value.
      */
-    protected $network;
+    public readonly string $network;
 
     /**
      * @ignore
+     *
+     * @param array<string, mixed> $raw
      */
     public function __construct(array $raw)
     {
-        parent::__construct($raw);
-        $this->autonomousSystemNumber = $this->get('autonomous_system_number');
-        $this->autonomousSystemOrganization =
-            $this->get('autonomous_system_organization');
-        $ipAddress = $this->get('ip_address');
+        $this->autonomousSystemNumber = $raw['autonomous_system_number'] ?? null;
+        $this->autonomousSystemOrganization
+            = $raw['autonomous_system_organization'] ?? null;
+        $ipAddress = $raw['ip_address'];
         $this->ipAddress = $ipAddress;
-        $this->network = Util::cidr($ipAddress, $this->get('prefix_len'));
+        $this->network = Util::cidr($ipAddress, $raw['prefix_len']);
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    public function jsonSerialize(): ?array
+    {
+        $js = [];
+
+        if ($this->autonomousSystemNumber !== null) {
+            $js['autonomous_system_number'] = $this->autonomousSystemNumber;
+        }
+        if ($this->autonomousSystemOrganization !== null) {
+            $js['autonomous_system_organization'] = $this->autonomousSystemOrganization;
+        }
+        $js['ip_address'] = $this->ipAddress;
+        $js['network'] = $this->network;
+
+        return $js;
     }
 }
