@@ -8,43 +8,54 @@ use GeoIp2\Util;
 
 /**
  * This class provides the GeoIP2 Connection-Type model.
- *
- * @property-read string|null $connectionType The connection type may take the
- *     following values: "Dialup", "Cable/DSL", "Corporate", "Cellular".
- *     Additional values may be added in the future.
- * @property-read string $ipAddress The IP address that the data in the model is
- *     for.
- * @property-read string $network The network in CIDR notation associated with
- *      the record. In particular, this is the largest network where all of the
- *      fields besides $ipAddress have the same value.
  */
-class ConnectionType extends AbstractModel
+class ConnectionType implements \JsonSerializable
 {
     /**
-     * @var string|null
+     * @var string|null The connection type may take the
+     *                  following values: "Dialup", "Cable/DSL", "Corporate", "Cellular", and
+     *                  "Satellite". Additional values may be added in the future.
      */
-    protected $connectionType;
+    public readonly ?string $connectionType;
 
     /**
-     * @var string
+     * @var string the IP address that the data in the model is
+     *             for
      */
-    protected $ipAddress;
+    public readonly string $ipAddress;
 
     /**
-     * @var string
+     * @var string The network in CIDR notation associated with
+     *             the record. In particular, this is the largest network where all of the
+     *             fields besides $ipAddress have the same value.
      */
-    protected $network;
+    public readonly string $network;
 
     /**
      * @ignore
+     *
+     * @param array<string, mixed> $raw
      */
     public function __construct(array $raw)
     {
-        parent::__construct($raw);
-
-        $this->connectionType = $this->get('connection_type');
-        $ipAddress = $this->get('ip_address');
+        $this->connectionType = $raw['connection_type'] ?? null;
+        $ipAddress = $raw['ip_address'];
         $this->ipAddress = $ipAddress;
-        $this->network = Util::cidr($ipAddress, $this->get('prefix_len'));
+        $this->network = Util::cidr($ipAddress, $raw['prefix_len']);
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    public function jsonSerialize(): ?array
+    {
+        $js = [];
+        if ($this->connectionType !== null) {
+            $js['connection_type'] = $this->connectionType;
+        }
+        $js['ip_address'] = $this->ipAddress;
+        $js['network'] = $this->network;
+
+        return $js;
     }
 }
