@@ -8,43 +8,54 @@ use GeoIp2\Util;
 
 /**
  * This class provides the GeoIP2 Domain model.
- *
- * @property-read string|null $domain The second level domain associated with the
- *     IP address. This will be something like "example.com" or
- *     "example.co.uk", not "foo.example.com".
- * @property-read string $ipAddress The IP address that the data in the model is
- *     for.
- * @property-read string $network The network in CIDR notation associated with
- *      the record. In particular, this is the largest network where all of the
- *      fields besides $ipAddress have the same value.
  */
-class Domain extends AbstractModel
+class Domain implements \JsonSerializable
 {
     /**
-     * @var string|null
+     * @var string|null The second level domain associated with the
+     *                  IP address. This will be something like "example.com" or
+     *                  "example.co.uk", not "foo.example.com".
      */
-    protected $domain;
+    public readonly ?string $domain;
 
     /**
-     * @var string
+     * @var string the IP address that the data in the model is
+     *             for
      */
-    protected $ipAddress;
+    public readonly string $ipAddress;
 
     /**
-     * @var string
+     * @var string The network in CIDR notation associated with
+     *             the record. In particular, this is the largest network where all of the
+     *             fields besides $ipAddress have the same value.
      */
-    protected $network;
+    public readonly string $network;
 
     /**
      * @ignore
+     *
+     * @param array<string, mixed> $raw
      */
     public function __construct(array $raw)
     {
-        parent::__construct($raw);
-
-        $this->domain = $this->get('domain');
-        $ipAddress = $this->get('ip_address');
+        $this->domain = $raw['domain'] ?? null;
+        $ipAddress = $raw['ip_address'];
         $this->ipAddress = $ipAddress;
-        $this->network = Util::cidr($ipAddress, $this->get('prefix_len'));
+        $this->network = Util::cidr($ipAddress, $raw['prefix_len']);
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    public function jsonSerialize(): ?array
+    {
+        $js = [];
+        if ($this->domain !== null) {
+            $js['domain'] = $this->domain;
+        }
+        $js['ip_address'] = $this->ipAddress;
+        $js['network'] = $this->network;
+
+        return $js;
     }
 }
