@@ -7,6 +7,5 @@ namespace GeoIp2\Exception;
 /**
  * This class represents a generic error.
  */
-class AuthenticationException extends GeoIp2Exception
-{
-}
+// phpcs:disable
+class AuthenticationException extends GeoIp2Exception {}
