@@ -7,6 +7,5 @@ namespace GeoIp2\Exception;
 /**
  * This class represents a generic error.
  */
-class GeoIp2Exception extends \Exception
-{
-}
+// phpcs:disable
+class GeoIp2Exception extends \Exception {}
