@@ -8,16 +8,35 @@ namespace GeoIp2\Record;
  * Contains data about your account.
  *
  * This record is returned by all location services and databases.
- *
- * @property-read int|null $queriesRemaining The number of remaining queries you
- * have for the service you are calling.
  */
-class MaxMind extends AbstractRecord
+class MaxMind implements \JsonSerializable
 {
+    /**
+     * @var int|null the number of remaining queries you
+     *               have for the service you are calling
+     */
+    public readonly ?int $queriesRemaining;
+
     /**
      * @ignore
      *
-     * @var array<string>
+     * @param array<string, mixed> $record
      */
-    protected $validAttributes = ['queriesRemaining'];
+    public function __construct(array $record)
+    {
+        $this->queriesRemaining = $record['queries_remaining'] ?? null;
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    public function jsonSerialize(): array
+    {
+        $js = [];
+        if ($this->queriesRemaining !== null) {
+            $js['queries_remaining'] = $this->queriesRemaining;
+        }
+
+        return $js;
+    }
 }
