@@ -10,24 +10,39 @@ namespace GeoIp2\Record;
  * This class contains the country-level data associated with an IP address
  * for the IP's represented country. The represented country is the country
  * represented by something like a military base.
- *
- * @property-read string|null $type A string indicating the type of entity that is
- * representing the country. Currently we only return <code>military</code>
- * but this could expand to include other types in the future.
  */
 class RepresentedCountry extends Country
 {
     /**
+     * @var string|null A string indicating the type of entity that is
+     *                  representing the country. Currently we only return <code>military</code>
+     *                  but this could expand to include other types in the future.
+     */
+    public readonly ?string $type;
+
+    /**
      * @ignore
      *
-     * @var array<string>
+     * @param array<string, mixed> $record
+     * @param list<string>         $locales
      */
-    protected $validAttributes = [
-        'confidence',
-        'geonameId',
-        'isInEuropeanUnion',
-        'isoCode',
-        'names',
-        'type',
-    ];
+    public function __construct(array $record, array $locales = ['en'])
+    {
+        parent::__construct($record, $locales);
+
+        $this->type = $record['type'] ?? null;
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    public function jsonSerialize(): array
+    {
+        $js = parent::jsonSerialize();
+        if ($this->type !== null) {
+            $js['type'] = $this->type;
+        }
+
+        return $js;
+    }
 }
